backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "news agregator's backend service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.7",
    "aiobotocore>=2.24.3",
    "aiohttp>=3.12.15",
    "fastapi[standard]>=0.116.1",
    "filetype>=1.2.0",
    "pika>=1.3.2",
    "pika-stubs>=0.1.3",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
]
